Post-It-Sticker Implementation
Was added 3 factory component (ElementContainer, ModalContainer and StickerContainer).
The first component is the most generic class to create containers
The second it's an series of method to create the modal's container, included two types at moment (add sticker and remove sticker)
The third component will create the main content of the Stickers.

Also was added the Controller component, however, it is being tested and will be fully implemented as a Controller class.